Remove AI.
adimittedly, i re-wrote one or two sentences of the opinion with ai.
i *really* need to be careful with my wording here, i do not want to make it seem like i am rooting for terrorist sites being spread on the internet. i reverted this change though as it seems to unnatural.